Tox: porting to a new api pt3
git-svn-id: http://svn.miranda-ng.org/main/branches/unsane@12531 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ tox/src/tox_transfer.h
@@ -18,17 +18,17 @@ struct FileTransferParam
 	PROTOFILETRANSFERSTATUS pfts;
 	FILE_TRANSFER_STATUS status;
 	FILE *hFile;
-	int32_t friendNumber;
-	uint8_t fileNumber;
-	int64_t transferNumber;
+	uint32_t friendNumber;
+	uint32_t fileNumber;
+	uint64_t transferNumber;
 
-	FileTransferParam(int32_t friendNumber, uint8_t fileNumber, const TCHAR *fileName, uint64_t fileSize)
+	FileTransferParam(uint32_t friendNumber, uint32_t fileNumber, const TCHAR *fileName, uint64_t fileSize)
 	{
 		status = NONE;
 		hFile = NULL;
 		this->friendNumber = friendNumber;
 		this->fileNumber = fileNumber;
-		transferNumber = (((int64_t)friendNumber) << 32) | ((int64_t)fileNumber);
+		transferNumber = (((uint32_t)friendNumber) << 32) | ((uint32_t)fileNumber);
 
 		pfts.cbSize = sizeof(PROTOFILETRANSFERSTATUS);
 		pfts.flags = PFTS_TCHAR;
